HW: Change u8 pad specifier params to int where applicable
Using u8 as indexers is kind of silly, since the rest of the public API
essentially uses int for this sort of thing. Changing these to int also
gets rid of quite a few implicit truncations.

This also allows for getting rid of similar silliness in the netplay API.

+++ Source/Core/Core/HW/GCKeyboard.h
@@ -17,5 +17,5 @@ void LoadConfig();
 
 InputConfig* GetConfig();
 
-KeyboardStatus GetStatus(u8 port);
+KeyboardStatus GetStatus(int port);
 }
